+++ docker/scripts/load_data.py
@@ -14,6 +14,7 @@
 
 import sys
 import time
+from typing import Union
 
 import argilla as rg
 import pandas as pd
@@ -162,44 +163,41 @@ class LoadDatasets:
         dataset.push_to_argilla(name=repo_id.split("/")[-1])
 
     @staticmethod
-    def load_mock_dataset_for_filter_metadata():
-        print("Loading Mock dataset for metadata properties test")
-        import random
+    def build_error_analysis_record(
+        row: pd.Series, legacy: bool = False
+    ) -> Union[rg.FeedbackRecord, rg.TextClassificationRecord]:
+        fields = {
+            "user-message-1": row["HumanMessage1"],
+            "llm-output": row["llm_output"]
+            if not row["llm_output"].__contains__("```json")
+            else row["llm_output"].replace("'", '"'),
+            "ai-message": (f"```json\n{row['AIMessage']}\n```" if not legacy else row["AIMessage"]).replace("'", '"'),
+            "function-message": (f"```json\n{row['FunctionMessage']}\n```" if not legacy else row["AIMessage"]).replace(
+                "'", '"'
+            ),
+            "system-message": "You are an AI assistant name ACME",
+            "langsmith-url": f"https://smith.langchain.com/o/{row['parent_id']}",
+        }
+        metadata = {
+            "correctness-langsmith": row["correctness_langsmith"],
+            "model-name": row["model_name"],
+            "temperature": row["temperature"],
+            "max-tokens": row["max_tokens"],
+            "cpu-user": row["cpu_time_user"],
+            "cpu-system": row["cpu_time_system"],
+            "library-version": row["library_version"],
+        }
 
-        import argilla as rg
-
-        dataset = rg.FeedbackDataset.from_huggingface("argilla/oasst_response_quality", split="train")
-
-        records = dataset.records[:400]
-
-        for record in records:
-            record.metadata.update(
-                {
-                    "terms-prop": random.choice("abc"),
-                    "int-prop": random.randint(-10, 100),
-                    "float-prop": random.uniform(0, 100),
-                }
+        if legacy:
+            return rg.TextClassificationRecord(
+                inputs=fields, metadata=metadata, vectors=eval(row["vectors"]), multi_label=True
             )
-        meta_ds = rg.FeedbackDataset(
-            fields=dataset.fields,
-            questions=dataset.questions,
-            metadata_properties=[
-                rg.TermsMetadataProperty(name="terms-prop", values=["a", "b", "c"]),
-                rg.IntegerMetadataProperty(name="int-prop", min=-10, max=100),
-                rg.FloatMetadataProperty(name="float-prop", min=0.0, max=100),
-            ],
-        )
-
-        meta_ds.add_records(records)
-        meta_ds.push_to_argilla("dataset-with-metadata-properties")
+        return rg.FeedbackRecord(fields=fields, metadata=metadata)
 
     @staticmethod
-    def load_meta7():
-        print("Loading Meta 7 dataset")
-        df = pd.read_csv(
-            "https://raw.githubusercontent.com/argilla-io/dataset_examples/main/synthetic_data_1000_records.csv"
-        )
-        df["llm_output"] = df["llm_output"].str.replace("'", '"')
+    def load_error_analysis():
+        print("Loading Error Analysis dataset as a `FeedbackDataset` (Alpha)")
+        df = pd.read_csv("https://raw.githubusercontent.com/argilla-io/dataset_examples/main/synthetic_data_v2.csv")
 
         fields = [
             rg.TextField(name="user-message-1", use_markdown=True),
@@ -223,40 +221,29 @@ class LoadDatasets:
             rg.TermsMetadataProperty(name="correctness-langsmith", values=df.correctness_langsmith.unique().tolist()),
             rg.TermsMetadataProperty(name="model-name", values=df.model_name.unique().tolist()),
             rg.FloatMetadataProperty(name="temperature", min=df.temperature.min(), max=df.temperature.max()),
-            rg.FloatMetadataProperty(name="max-tokens", min=df.max_tokens.min(), max=df.max_tokens.max()),
+            rg.IntegerMetadataProperty(name="max-tokens", min=df.max_tokens.min(), max=df.max_tokens.max()),
             rg.FloatMetadataProperty(name="cpu-user", min=df.cpu_time_user.min(), max=df.cpu_time_user.max()),
             rg.FloatMetadataProperty(name="cpu-system", min=df.cpu_time_system.min(), max=df.cpu_time_system.max()),
             rg.TermsMetadataProperty(name="library-version", values=df.library_version.unique().tolist()),
         ]
 
-        def build_record(row):
-            fields = {
-                "user-message-1": row["HumanMessage1"],
-                "llm-output": f"```json\n{row.llm_output}\n```",
-                "ai-message": row["AIMessage"],
-                "function-message": row["FunctionMessage"],
-                "system-message": "You are an AI assistant name ACME",
-                "langsmith-url": f"https://smith.langchain.com/o/{row.parent_id}",
-            }
-            metadata = {
-                "correctness-langsmith": row["correctness_langsmith"],
-                "model-name": row["model_name"],
-                "temperature": row["temperature"],
-                "max-tokens": row["max_tokens"],
-                "cpu-user": row["cpu_time_user"],
-                "cpu-system": row["cpu_time_system"],
-                "library-version": row["library_version"],
-            }
-
-            return rg.FeedbackRecord(fields=fields, metadata=metadata)
-
-        records = [build_record(r) for i, r in df.iterrows()]
-
         dataset = rg.FeedbackDataset(fields=fields, questions=questions, metadata_properties=metadata)
+        dataset.add_records(records=[LoadDatasets.build_error_analysis_record(row) for _, row in df.iterrows()])
+        dataset.push_to_argilla(name="error-analysis-with-feedback-alpha")
 
-        dataset.add_records(records[0:100])
+    @staticmethod
+    def load_error_analysis_textcat_version():
+        print("Loading Error Analysis dataset as a `DatasetForTextClassification`")
+        df = pd.read_csv(
+            "https://raw.githubusercontent.com/argilla-io/dataset_examples/main/synthetic_data_v2_with_vectors.csv"
+        )
 
-        dataset.push_to_argilla(name="meta7", workspace="admin")
+        labels = ["follow-up needed", "reviewed", "no-repro", "not-helpful", "empty-response", "critical"]
+        settings = rg.TextClassificationSettings(label_schema=labels)
+        rg.configure_dataset_settings(name="error-analysis-with-text-classification", settings=settings)
+
+        records = [LoadDatasets.build_error_analysis_record(row, legacy=True) for _, row in df.iterrows()]
+        rg.log(name="error-analysis-with-text-classification", records=records, batch_size=25)
 
 
 if __name__ == "__main__":
@@ -272,10 +259,8 @@ if __name__ == "__main__":
                 response = requests.get("http://0.0.0.0:6900/")
                 if response.status_code == 200:
                     ld = LoadDatasets(API_KEY)
-                    # TODO(@frascuchon): Remove these datasets creation
-                    ld.load_mock_dataset_for_filter_metadata()
-                    ld.load_meta7()
-                    # END TODO
+                    ld.load_error_analysis()
+                    ld.load_error_analysis_textcat_version()
                     ld.load_feedback_dataset_from_huggingface(
                         repo_id="argilla/databricks-dolly-15k-curated-en",
                         split="train",
@@ -301,7 +286,6 @@ if __name__ == "__main__":
                             split="train",
                             samples=100,
                         )
-
             except requests.exceptions.ConnectionError:
                 pass
             except Exception as e: