+++ frontend/v1/infrastructure/repositories/RecordRepository.ts
@@ -5,7 +5,8 @@ import {
   BackendAnswerCombinations,
   BackendResponse,
   BackendRecordStatus,
-  Response,
+  BackendSearchRecords,
+  ResponseWithTotal,
 } from "../types";
 import { RecordAnswer } from "@/v1/domain/entities/record/RecordAnswer";
 import { Record } from "@/v1/domain/entities/record/Record";
@@ -27,7 +28,8 @@ export class RecordRepository {
     howMany: number,
     status: string,
     searchText: string,
-    metadata: string[]
+    metadata: string[],
+    sortBy: string[]
   ): Promise<BackedRecords> {
     if (searchText?.length)
       return this.getRecordsByText(
@@ -36,7 +38,8 @@ export class RecordRepository {
         howMany,
         status,
         searchText,
-        metadata
+        metadata,
+        sortBy
       );
 
     return this.getRecordsDatasetId(
@@ -44,7 +47,8 @@ export class RecordRepository {
       fromRecord,
       howMany,
       status,
-      metadata
+      metadata,
+      sortBy
     );
   }
 
@@ -128,20 +132,31 @@ export class RecordRepository {
     fromRecord: number,
     howMany: number,
     status: string,
-    metadata: string[]
+    metadata: string[],
+    sortBy: string[]
   ): Promise<BackedRecords> {
     try {
       const url = `/v1/me/datasets/${datasetId}/records`;
 
-      const params = this.createParams(fromRecord, howMany, status, metadata);
+      const params = this.createParams(
+        fromRecord,
+        howMany,
+        status,
+        metadata,
+        sortBy
+      );
 
-      const { data } = await this.axios.get<Response<BackedRecord[]>>(url, {
-        params,
-      });
+      const { data } = await this.axios.get<ResponseWithTotal<BackedRecord[]>>(
+        url,
+        {
+          params,
+        }
+      );
+      const { items: records, total } = data;
 
       return {
-        records: data.items,
-        total: data.items.length,
+        records,
+        total,
       };
     } catch (err) {
       throw {
@@ -156,7 +171,8 @@ export class RecordRepository {
     howMany: number,
     status: string,
     searchText: string,
-    metadata: string[]
+    metadata: string[],
+    sortBy: string[]
   ): Promise<BackedRecords> {
     try {
       const url = `/v1/me/datasets/${datasetId}/records/search`;
@@ -171,9 +187,17 @@ export class RecordRepository {
         })
       );
 
-      const params = this.createParams(fromRecord, howMany, status, metadata);
+      const params = this.createParams(
+        fromRecord,
+        howMany,
+        status,
+        metadata,
+        sortBy
+      );
 
-      const { data } = await this.axios.post(url, body, { params });
+      const { data } = await this.axios.post<
+        ResponseWithTotal<BackendSearchRecords[]>
+      >(url, body, { params });
 
       const { items, total } = data;
 
@@ -215,7 +239,8 @@ export class RecordRepository {
     fromRecord: number,
     howMany: number,
     status: string,
-    metadata: string[]
+    metadata: string[],
+    sortBy: string[]
   ) {
     const offset = `${fromRecord - 1}`;
     const backendStatus = status === "pending" ? "missing" : status;
@@ -231,6 +256,10 @@ export class RecordRepository {
 
     metadata?.forEach((query) => {
       params.append("metadata", query);
+    });
+
+    sortBy?.forEach((sort) => {
+      params.append("sort_by", sort);
     });
 
     return params;