+++ src/argilla/server/enums.py
@@ -65,3 +65,13 @@ class MetadataPropertyType(str, Enum):
     terms = "terms"  # Textual types with a fixed value list
     integer = "integer"  # Integer values
     float = "float"  # Decimal values
+
+
+class RecordSortField(str, Enum):
+    inserted_at = "inserted_at"
+    updated_at = "updated_at"
+
+
+class SortOrder(str, Enum):
+    asc = "asc"
+    desc = "desc"