+++ src/argilla/server/search_engine/base.py
@@ -13,15 +13,14 @@
 #  limitations under the License.
 
 import dataclasses
-import re
 from abc import ABCMeta, abstractmethod
 from contextlib import asynccontextmanager
-from typing import Any, AsyncGenerator, Dict, Iterable, List, Literal, Optional, Union
+from typing import AsyncGenerator, Iterable, List, Literal, Optional, Union
 from uuid import UUID
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
-from argilla.server.enums import ResponseStatus, ResponseStatusFilter
+from argilla.server.enums import RecordSortField, ResponseStatusFilter, SortOrder
 from argilla.server.models import Dataset, MetadataProperty, Record, Response, User
 
 __all__ = [
@@ -122,8 +121,8 @@ class SearchResponses:
 
 
 class SortBy(BaseModel):
-    field: Union[MetadataProperty, Literal["inserted_at"], Literal["updated_at"]]
-    order: Union[Literal["asc"], Literal["desc"]] = "asc"
+    field: Union[MetadataProperty, RecordSortField]
+    order: SortOrder = SortOrder.asc
 
     class Config:
         arbitrary_types_allowed = True
@@ -208,6 +207,6 @@ class SearchEngine(metaclass=ABCMeta):
         metadata_filters: Optional[List[MetadataFilter]] = None,
         offset: int = 0,
         limit: int = 100,
-        sort_by: List[SortBy] = None,
+        sort_by: Optional[List[SortBy]] = None,
     ) -> SearchResponses:
         pass